+++ src/deluge/gui/menu_item/note/iterance_preset.h
