+++ src/deluge/gui/ui/menus.cpp
@@ -102,13 +102,13 @@
 #include "gui/menu_item/monitor/mode.h"
 #include "gui/menu_item/mpe/direction_selector.h"
 #include "gui/menu_item/note/fill.h"
-#include "gui/menu_item/note/iterance.h"
+#include "gui/menu_item/note/iterance_preset.h"
 #include "gui/menu_item/note/iterance_divisor.h"
 #include "gui/menu_item/note/iterance_step_toggle.h"
 #include "gui/menu_item/note/probability.h"
 #include "gui/menu_item/note/velocity.h"
 #include "gui/menu_item/note_row/fill.h"
-#include "gui/menu_item/note_row/iterance.h"
+#include "gui/menu_item/note_row/iterance_preset.h"
 #include "gui/menu_item/note_row/iterance_divisor.h"
 #include "gui/menu_item/note_row/iterance_step_toggle.h"
 #include "gui/menu_item/note_row/probability.h"
