+++ src/deluge/gui/ui/menus.h
@@ -4,12 +4,12 @@
 #include "gui/menu_item/defaults/swing_interval.h"
 #include "gui/menu_item/firmware/version.h"
 #include "gui/menu_item/note/fill.h"
-#include "gui/menu_item/note/iterance.h"
+#include "gui/menu_item/note/iterance_preset.h"
 #include "gui/menu_item/note/iterance_divisor.h"
 #include "gui/menu_item/note/iterance_step_toggle.h"
 #include "gui/menu_item/note/probability.h"
 #include "gui/menu_item/note_row/fill.h"
-#include "gui/menu_item/note_row/iterance.h"
+#include "gui/menu_item/note_row/iterance_preset.h"
 #include "gui/menu_item/note_row/iterance_divisor.h"
 #include "gui/menu_item/note_row/iterance_step_toggle.h"
 #include "gui/menu_item/note_row/probability.h"
